pyproject.toml
[tool.poetry]
name = "gcp_microservice_management"
version = "0.3.0"
description = "A package to deploy microservices on GCP"
authors = ["David Piper <dbpiper@gmail.com>"]
packages = [{ include = "gcp_microservice_management" }]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-run = "*"
google-cloud-api-gateway = "*"
google-api-core = "*"
google-auth = "*"
google-auth-oauthlib = "*"
docker = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
